Move CompleteType, getCompletionUrl(), and AutoCompleteAction out of specimen, back into study
