+++ api/src/org/labkey/api/study/query/PublishResultsQueryView.java
@@ -56,6 +56,7 @@ import org.labkey.api.study.ParticipantVisit;
 import org.labkey.api.study.SpecimenService;
 import org.labkey.api.study.Study;
 import org.labkey.api.study.StudyService;
+import org.labkey.api.study.StudyUrls;
 import org.labkey.api.study.TimepointType;
 import org.labkey.api.study.Visit;
 import org.labkey.api.study.actions.StudyPickerColumn;
@@ -822,7 +823,7 @@ public class PublishResultsQueryView extends ResultsQueryView
         protected ActionURL getCompletionBase(RenderContext ctx)
         {
             Container c = rowTargetStudy(_resolverHelper, ctx);
-            return SpecimenService.get().getCompletionURL(c, CompletionType.ParticipantId);
+            return PageFlowUtil.urlProvider(StudyUrls.class).getCompletionURL(c, CompletionType.ParticipantId);
         }
 
         @Override
@@ -844,7 +845,7 @@ public class PublishResultsQueryView extends ResultsQueryView
         protected ActionURL getCompletionBase(RenderContext ctx)
         {
             Container c = rowTargetStudy(_resolverHelper, ctx);
-            return SpecimenService.get().getCompletionURL(c, CompletionType.VisitId);
+            return PageFlowUtil.urlProvider(StudyUrls.class).getCompletionURL(c, CompletionType.VisitId);
         }
 
         @Override
