+++ study/src/org/labkey/study/controllers/specimen/SpecimenController.java
@@ -90,7 +90,6 @@ import org.labkey.api.specimen.SpecimenRequestManager.SpecimenRequestInput;
 import org.labkey.api.specimen.SpecimenRequestStatus;
 import org.labkey.api.specimen.SpecimenSearchWebPart;
 import org.labkey.api.specimen.Vial;
-import org.labkey.api.specimen.actions.AutoCompleteAction;
 import org.labkey.api.specimen.actions.ManageReqsBean;
 import org.labkey.api.specimen.actions.ReportConfigurationBean;
 import org.labkey.api.specimen.actions.SpecimenReportActions;
@@ -230,7 +229,6 @@ public class SpecimenController extends BaseStudyController
         ShowUploadSpecimensAction.ImportCompleteAction.class,
         ShowGroupMembersAction.class,
         ShowSearchAction.class,
-        AutoCompleteAction.class,
         ParticipantCommentAction.SpecimenCommentInsertAction.class,
         ParticipantCommentAction.SpecimenCommentUpdateAction.class,
